Fix compiler warnings for the x64 build.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@16486 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/SVN/SVNFolderStatus.cpp
@@ -285,7 +285,7 @@ const FileStatusCacheEntry * SVNFolderStatus::BuildCache(const CTSVNPath& filepa
 DWORD SVNFolderStatus::GetTimeoutValue()
 {
 	DWORD timeout = SVNFOLDERSTATUS_CACHETIMEOUT;
-	DWORD factor = m_cache.size()/200;
+	DWORD factor = (DWORD)m_cache.size()/200;
 	if (factor==0)
 		factor = 1;
 	return factor*timeout;
