+++ src/Utils/MiscUI/HistoryCombo.cpp
@@ -1,6 +1,6 @@
 // TortoiseSVN - a Windows shell extension for easy version control
 
-// Copyright (C) 2003-2008 - TortoiseSVN
+// Copyright (C) 2003-2009 - TortoiseSVN
 
 // This program is free software; you can redistribute it and/or
 // modify it under the terms of the GNU General Public License
@@ -72,7 +72,7 @@ BOOL CHistoryCombo::PreTranslateMessage(MSG* pMsg)
 			CPoint pt;
 			pt.x = LOWORD(pMsg->lParam);
 			pt.y = HIWORD(pMsg->lParam);
-			OnMouseMove(pMsg->wParam, pt);
+			OnMouseMove((UINT)pMsg->wParam, pt);
 			return TRUE;
 		}
 	}
